feat: initial commit for the monorepo structuring
I've decided to stick with `just` for command management, as they
provide a module structure that's very favorable for monorepos.

each package is going to be stored at `src/{package}`, with its own
`pyproject.toml` and `mod.just` files, for defining relevant commands,
tests, lints, and so on.

the main `justfile` in the workspace root then serves as the
coordinator, calling all the subfiles in order to create commands to
run in the CI and pre-commit.