feat(postgrest): make mypy accept the type definitions
the generics usage is completely wrong, and had to be entirely
removed. the idea was to make the type use the generic to know how to
validate the payload, but that does not work in python, as type hints
cannot contribute with computational content, and there is no
monomorphization of the calls.

instead, we use a simple JSON type that reflects the fact that the
best information we can know about the result of a query is that it
returns valid json, and nothing more.